requirements.txt
PyJWT
requests
requests_oauthlib